fix: address review feedback — nullable userId exclusion and test deduplication
